+++ src/core/opengl/render_gl.h
@@ -751,7 +751,7 @@ public:
                               int16_t rotateAngle = 0, Point_t *center = nullptr, unsigned int flip = X_FLIP_NONE,
                               XTColor color = XTColor()) override;
 
-    void renderTextureScale(double xDst, double yDst, double wDst, double hDst,
+    void renderTextureScale(int xDst, int yDst, int wDst, int hDst,
                             StdPicture &tx,
                             XTColor color = XTColor()) override;
